# app/Http/Controllers/Api/EstampilladoController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Controller;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Validator;
use App\Helpers\QRHelper;
class EstampilladoController extends Controller
{
    /**
     * Estampillar.
     *
     * @OA\Post(
     *    path="/api/estampillar",
     *    summary="Estampillar",
     *    description="Aprobar/Rechazar una tarea asignada, del empleado autenticado",
     *    operationId="estampillar",
     *    tags={"estampillar"},
     *    security={{ "apiAuth": {} }},
     *    @OA\RequestBody(
     *       @OA\MediaType(
     *          mediaType="multipart/form-data",
     *          @OA\Schema(
     *             type="object",
     *             @OA\Property(
     *                property="frontend_endpoint",
     *                type="string"
     *             ),
     *             @OA\Property(
     *                property="numero_expediente",
     *                type="string"
     *             ),
     *             @OA\Property(
     *                property="archivo_estatuto",
     *                type="file"
     *             ),
     *          ),
     *      )
     *    ),
     *    @OA\Response(
     *       response=200,
     *       description="Estampillado exitoso."
     *    ),
     *     @OA\Response(
     *       response=401,
     *       description="Unauthorized"
     *    ),
     *    @OA\Response(
     *       response=500,
     *       description="500 internal server error",
     *       @OA\JsonContent(
     *          example="500 internal server error"
     *       )
     *    ),
     * ) 
     * 
     * @param  \Illuminate\Http\Request $request
     * @return \Illuminate\Http\JsonResponse
     */
    public function estampillar(Request $request)
    {
        $validator = Validator::make($request->all(), [
            'numero_expediente' => 'required|string|between:2,100',
            //'archivo_estatuto' => 'required|mimes:docx,odt,pdf'
        ]);

        if ($validator->fails()) {
            $errors = $validator->errors();
            return response()->json($errors, 400);
        }

        $numeroHash = "123";

        // Generar Código QR
        $qrHelper = new QRHelper();
        $qr = $qrHelper->generarQR("{$request->input('frontend_endpoint')}/{$numeroHash}");

        return response()->json([
            "numero_hash" => $numeroHash,
            "qr" => $qr,
        ], 200);
    }
}