# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "dssd-unlp-2021-grupo11-estampillado-service",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "description": "Login con email y password",
                "operationId": "authLogin",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succesful login",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "401 Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Unauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "description": "Logout",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/estampillar": {
            "post": {
                "tags": [
                    "estampillar"
                ],
                "summary": "Estampillar",
                "description": "Aprobar/Rechazar una tarea asignada, del empleado autenticado",
                "operationId": "estampillar",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "frontend_endpoint": {
                                        "type": "string"
                                    },
                                    "numero_expediente": {
                                        "type": "string"
                                    },
                                    "archivo_estatuto": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estampillado exitoso."
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Token Based",
                "name": "Token Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}